Respect download limits enforced by remote server, update ffsend-api
See issue timvisee/ffsend#17

+++ src/cmd/matcher/params.rs
@@ -26,7 +26,7 @@ impl<'a: 'b, 'b> ParamsMatcher<'a> {
     }
 
     /// Get the download limit.
-    pub fn download_limit(&'a self) -> Option<u8> {
+    pub fn download_limit(&'a self) -> Option<usize> {
         ArgDownloadLimit::value(self.matches)
     }
 }
